categories edit fixed at Admin panel

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\CategoryRequest;
 use App\Models\Category;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
@@ -36,11 +37,16 @@ class CategoryController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(CategoryRequest $request)
     {
-        $path = $request->file('img')->store('categories');
+
         $params = $request->all();
-        $params['img'] = $path;
+        unset($params['img']);
+
+        if ($request->has('img')) {
+            $path = $request->file('img')->store('categories');
+            $params['img'] = $path;
+        }
 
       Category::create($params);
       return redirect()->route('categories.index');
@@ -75,12 +81,18 @@ class CategoryController extends Controller
      * @param  \App\Models\Category  $category
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Category $category)
+    public function update(CategoryRequest $request, Category $category)
     {
-        Storage::delete($category->img);
-        $path = $request->file('img')->store('categories');
         $params = $request->all();
-        $params['img'] = $path;
+        unset($params['img']);
+
+        if($request->has('img')){
+            Storage::delete($category->img);
+            $path = $request->file('img')->store('categories');
+            $params['img'] = $path;
+        }
+
+
         $category->update($params);
         return redirect()->route('categories.index');
     }
